Refactoring and additions
Added call to set_RNG_LEDs to actually set the LEDs (probably forgot
this one because I was not able to test this).
Also added additional call to shutdown_modules after loss through
running out of lives. Refactored code accordingly.
Lastly, reduced the number of modules on internal module error (win
state) to ensure that the player is not punished for any module
implementation errors.