Added support for needy modules
A needy module is any kind of module that can not actually be solved but
instead require occasional attention by the player. In the game start
phase, a needy module sends a 2 to acknowledge the main controllers game
start request. The controller thus knows, that the module is needy and
does not count towards a full solution of the puzzle at hand. The code
changes in this commit have not yet been tested.